What a night...
We did a lot! (In theory - none of this has been tested)

We split everything up into nice classes and created an object constructor for the encoders.

We wrote the functions for calculating our degrees from the two X encoders.

We also can calculate the distance traveled if the robot cannot turn.

TO-DO
Finish the algorithm that actually does the main part of odometry.

Sorry for the incoherent commit. It's 1:12 AM. We have been working for 6 hours, and I cannot think straight.

Buh-bye now!